Count the bytes in one integer type, not two
std::int64_t is long on Linux and long long in JUCE, so std::max could not
deduce anything from one of each and both Linux jobs failed to build.

+++ magda/engine/io/AudioFileSink.cpp
@@ -54,7 +54,7 @@ class ReportingStream final : public juce::OutputStream {
 
     bool write(const void* data, std::size_t numBytes) override {
         const auto ok = note(file_->write(data, numBytes));
-        report_->bytes = std::max(report_->bytes, file_->getPosition());
+        report_->bytes = std::max(report_->bytes, static_cast<std::int64_t>(file_->getPosition()));
         return ok;
     }
